cloud/device-service/pyproject.toml
[project]
name = "seld-device-service"
version = "0.1.0"
description = "SELD Digital Twin — Device registry & management gRPC service"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "protobuf>=5.29.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "structlog>=24.4.0",
    "cryptography>=44.0.0",
    "seld-common",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[tool.hatch.build.targets.wheel]
packages = ["src/device_service"]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "pytest-asyncio>=0.25.0", "ruff>=0.8.0"]

[tool.uv.sources]
seld-common = { path = "../../libs/seld-common", editable = true }
